[common] Split configuration parsing to multiple files.
