+++ src/common/Makefile.am
@@ -12,7 +12,7 @@ AM_CFLAGS = \
 
 modlibexec_LTLIBRARIES = libdroid-util.la
 
-libdroid_util_la_SOURCES = droid-util.c droid-util.h
+libdroid_util_la_SOURCES = droid-util.c droid-config.c conversion.c config-parser-legacy.c
 libdroid_util_la_LDFLAGS = -avoid-version -Wl,-z,noexecstack -lhybris-common
 libdroid_util_la_LIBADD = $(AM_LIBADD)
 libdroid_util_la_CFLAGS = $(AM_CFLAGS)
