+++ src/common/droid-util.c
@@ -58,7 +58,10 @@
 #include <pulsecore/strlist.h>
 #include <pulsecore/atomic.h>
 
+#include "version.h"
 #include "droid-util.h"
+#include "droid-config.h"
+#include "conversion.h"
 
 struct droid_quirk {
     const char *name;
@@ -75,848 +78,17 @@ struct droid_quirk valid_quirks[] = {
     { "realcall",               QUIRK_REALCALL              },
 };
 
-#define SLLIST_APPEND(t, head, item)                                \
-    do {                                                            \
-        item->next = NULL;                                          \
-        if (!head) {                                                \
-            head = item;                                            \
-        } else {                                                    \
-            t *_list;                                               \
-            for (_list = head; _list->next; _list = _list->next);   \
-            _list->next = item;                                     \
-        }                                                           \
-    } while (0)
-
-#define SLLIST_FOREACH(i, head)                                     \
-    for (i = (head); i; i = i->next)
-
-#define SLLIST_STEAL_FIRST(i, head)                                 \
-    do {                                                            \
-        if (head) {                                                 \
-            i = head;                                               \
-            head = head->next;                                      \
-        } else                                                      \
-            i = NULL;                                               \
-    } while (0)
-
-
-#define CONVERT_FUNC(TABL) \
-bool pa_convert_ ## TABL (uint32_t value, pa_conversion_field_t field, uint32_t *to_value) {                    \
-    for (unsigned int i = 0; i < sizeof( conversion_table_ ## TABL )/(sizeof(uint32_t)*2); i++) {               \
-        if ( conversion_table_ ## TABL [i][field] == value) {                                                   \
-            *to_value = conversion_table_ ## TABL [i][!field];                                                  \
-            return true;                                                                                        \
-        }                                                                                                       \
-    }                                                                                                           \
-    return false;                                                                                               \
-} struct __funny_extra_to_allow_semicolon
-
-/* Creates convert_format convert_channel etc.
- * bool pa_convert_func(uint32_t value, pa_conversion_field_t field, uint32_t *to_value);
- * return true if conversion succesful */
-CONVERT_FUNC(format);
-CONVERT_FUNC(output_channel);
-CONVERT_FUNC(input_channel);
 
 #define DEFAULT_PRIORITY (100)
-
-/* Section defining custom global configuration variables. */
-#define GLOBAL_CONFIG_EXT_TAG "custom_properties"
-
-/* From recent audio_policy_conf.h */
-#ifndef AUDIO_HAL_VERSION_TAG
-#define AUDIO_HAL_VERSION_TAG "audio_hal_version"
-#endif
-#ifndef GAINS_TAG
-#define GAINS_TAG "gains"
-#endif
-
-#define GAIN_TAG_PREFIX "gain_"
 
 
 static const char * const droid_combined_auto_outputs[3]    = { "primary", "low_latency", NULL };
 static const char * const droid_combined_auto_inputs[2]     = { "primary", NULL };
 
-static void droid_config_free(pa_droid_config_audio *config);
 static void droid_port_free(pa_droid_port *p);
 
 static pa_droid_stream *get_primary_output(pa_droid_hw_module *hw);
 static int input_stream_set_route(pa_droid_stream *s, audio_devices_t device);
-
-static bool string_convert_num_to_str(const struct string_conversion *list, const uint32_t value, const char **to_str) {
-    pa_assert(list);
-    pa_assert(to_str);
-
-    for (unsigned int i = 0; list[i].str; i++) {
-        if (list[i].value == value) {
-            *to_str = list[i].str;
-            return true;
-        }
-    }
-    return false;
-}
-
-static bool string_convert_str_to_num(const struct string_conversion *list, const char *str, uint32_t *to_value) {
-    pa_assert(list);
-    pa_assert(str);
-    pa_assert(to_value);
-
-    for (unsigned int i = 0; list[i].str; i++) {
-        if (pa_streq(list[i].str, str)) {
-            *to_value = list[i].value;
-            return true;
-        }
-    }
-    return false;
-}
-
-static char *list_string(struct string_conversion *list, uint32_t flags) {
-    char *str = NULL;
-    char *tmp;
-
-#if AUDIO_API_VERSION_MAJ >= 2
-    if (flags & AUDIO_DEVICE_BIT_IN)
-        flags &= ~AUDIO_DEVICE_BIT_IN;
-#endif
-
-    for (unsigned int i = 0; list[i].str; i++) {
-#if AUDIO_API_VERSION_MAJ >= 2
-        if (list[i].value & AUDIO_DEVICE_BIT_IN) {
-            if (popcount(list[i].value & ~AUDIO_DEVICE_BIT_IN) != 1)
-                continue;
-        } else
-#endif
-        if (popcount(list[i].value) != 1)
-            continue;
-
-        if (flags & list[i].value) {
-            if (str) {
-                tmp = pa_sprintf_malloc("%s|%s", str, list[i].str);
-                pa_xfree(str);
-                str = tmp;
-            } else {
-                str = pa_sprintf_malloc("%s", list[i].str);
-            }
-        }
-    }
-
-    return str;
-}
-
-
-/* Output device */
-bool pa_string_convert_output_device_num_to_str(audio_devices_t value, const char **to_str) {
-    return string_convert_num_to_str(string_conversion_table_output_device, (uint32_t) value, to_str);
-}
-
-bool pa_string_convert_output_device_str_to_num(const char *str, audio_devices_t *to_value) {
-    return string_convert_str_to_num(string_conversion_table_output_device, str, (uint32_t*) to_value);
-}
-
-char *pa_list_string_output_device(audio_devices_t devices) {
-    return list_string(string_conversion_table_output_device, devices);
-}
-
-/* Input device */
-bool pa_string_convert_input_device_num_to_str(audio_devices_t value, const char **to_str) {
-    return string_convert_num_to_str(string_conversion_table_input_device, (uint32_t) value, to_str);
-}
-
-bool pa_string_convert_input_device_str_to_num(const char *str, audio_devices_t *to_value) {
-    return string_convert_str_to_num(string_conversion_table_input_device, str, (uint32_t*) to_value);
-}
-
-char *pa_list_string_input_device(audio_devices_t devices) {
-    return list_string(string_conversion_table_input_device, devices);
-}
-
-/* Flags */
-bool pa_string_convert_flag_num_to_str(audio_output_flags_t value, const char **to_str) {
-    return string_convert_num_to_str(string_conversion_table_output_flag, (uint32_t) value, to_str);
-}
-
-bool pa_string_convert_flag_str_to_num(const char *str, audio_output_flags_t *to_value) {
-    return string_convert_str_to_num(string_conversion_table_output_flag, str, (uint32_t*) to_value);
-}
-
-char *pa_list_string_flags(audio_output_flags_t flags) {
-    return list_string(string_conversion_table_output_flag, flags);
-}
-
-bool pa_input_device_default_audio_source(audio_devices_t input_device, audio_source_t *default_source)
-{
-#if AUDIO_API_VERSION_MAJ >= 2
-    input_device &= ~AUDIO_DEVICE_BIT_IN;
-#endif
-
-    /* Note converting HAL values to different HAL values! */
-    for (unsigned int i = 0; i < sizeof(conversion_table_default_audio_source) / (sizeof(uint32_t) * 2); i++) {
-        if (conversion_table_default_audio_source[i][0] & input_device) {
-            *default_source = conversion_table_default_audio_source[i][1];
-            return true;
-        }
-    }
-    return false;
-}
-
-/* Config parser */
-
-#define WHITESPACE "\n\r \t"
-
-static int parse_list(const struct string_conversion *table,
-                      const char *str,
-                      uint32_t *dst,
-                      char **unknown_entries) {
-    int count = 0;
-    char *entry;
-    char *unknown = NULL;
-    const char *state = NULL;
-
-    pa_assert(table);
-    pa_assert(str);
-    pa_assert(dst);
-    pa_assert(unknown_entries);
-
-    *dst = 0;
-    *unknown_entries = NULL;
-
-    while ((entry = pa_split(str, "|", &state))) {
-        uint32_t d = 0;
-
-        if (!string_convert_str_to_num(table, entry, &d)) {
-            if (*unknown_entries) {
-                unknown = pa_sprintf_malloc("%s|%s", *unknown_entries, entry);
-                pa_xfree(*unknown_entries);
-                pa_xfree(entry);
-            } else
-                unknown = entry;
-
-            *unknown_entries = unknown;
-            continue;
-        }
-
-        *dst |= d;
-        count++;
-
-        pa_xfree(entry);
-    }
-
-    return count;
-}
-
-static bool parse_sampling_rates(const char *fn, const unsigned ln,
-                                 const char *str, uint32_t sampling_rates[32]) {
-    pa_assert(fn);
-    pa_assert(str);
-
-    char *entry;
-    const char *state = NULL;
-
-    uint32_t pos = 0;
-    while ((entry = pa_split(str, "|", &state))) {
-        int32_t val;
-
-#if AUDIO_API_VERSION_MAJ >= 3
-        if (pos == 0 && pa_streq(entry, "dynamic")) {
-            sampling_rates[pos++] = (uint32_t) -1;
-            pa_xfree(entry);
-            break;
-        }
-#endif
-
-        if (pos == AUDIO_MAX_SAMPLING_RATES) {
-            pa_log("[%s:%u] Too many sample rate entries (> %d)", fn, ln, AUDIO_MAX_SAMPLING_RATES);
-            pa_xfree(entry);
-            return false;
-        }
-
-        if (pa_atoi(entry, &val) < 0) {
-            pa_log("[%s:%u] Bad sample rate value %s", fn, ln, entry);
-            pa_xfree(entry);
-            return false;
-        }
-
-        sampling_rates[pos++] = val;
-
-        pa_xfree(entry);
-
-    }
-
-    sampling_rates[pos] = 0;
-
-    return true;
-}
-
-static bool check_and_log(const char *fn, const unsigned ln, const char *field,
-                          const int count, const char *str, char *unknown,
-                          const bool must_have_all) {
-    bool fail;
-
-    pa_assert(fn);
-    pa_assert(field);
-    pa_assert(str);
-
-    fail = must_have_all && unknown;
-
-    if (unknown) {
-        pa_log_info("[%s:%u] Unknown %s entries: %s", fn, ln, field, unknown);
-        pa_xfree(unknown);
-    }
-
-    if (count == 0 || fail) {
-        pa_log("[%s:%u] Failed to parse %s (%s).", fn, ln, field, str);
-        return false;
-    }
-
-    return true;
-}
-
-static bool parse_formats(const char *fn, const unsigned ln,
-                          const char *str, audio_format_t *formats) {
-    int count;
-    char *unknown = NULL;
-
-    pa_assert(fn);
-    pa_assert(str);
-    pa_assert(formats);
-
-#if AUDIO_API_VERSION_MAJ >= 3
-    /* Needs to be probed later */
-    if (pa_streq(str, "dynamic")) {
-        *formats = 0;
-        return true;
-    }
-#endif
-
-    count = parse_list(string_conversion_table_format, str, formats, &unknown);
-
-    return check_and_log(fn, ln, "formats", count, str, unknown, false);
-}
-
-static bool parse_channels(const char *fn, const unsigned ln,
-                           const char *str, bool in_output, audio_channel_mask_t *channels) {
-    int count;
-    char *unknown = NULL;
-
-    pa_assert(fn);
-    pa_assert(str);
-    pa_assert(channels);
-
-    /* Needs to be probed later */
-    if (pa_streq(str, "dynamic")) {
-        *channels = 0;
-        return true;
-    }
-
-    count = parse_list(in_output ? string_conversion_table_output_channels
-                                 : string_conversion_table_input_channels,
-                                 str, channels, &unknown);
-
-    return check_and_log(fn, ln, in_output ? "output channel_masks" : "input channel_masks",
-                         count, str, unknown, false);
-}
-
-static bool parse_devices(const char *fn, const unsigned ln,
-                          const char *str, bool in_output, audio_devices_t *devices, bool must_have_all) {
-    int count;
-    char *unknown = NULL;
-
-    pa_assert(fn);
-    pa_assert(str);
-    pa_assert(devices);
-
-    count = parse_list(in_output ? string_conversion_table_output_device
-                                 : string_conversion_table_input_device,
-                                 str, devices, &unknown);
-
-    return check_and_log(fn, ln, in_output ? "output devices" : "input devices",
-                         count, str, unknown, must_have_all);
-}
-
-static bool parse_output_flags(const char *fn, const unsigned ln,
-                        const char *str, audio_output_flags_t *flags) {
-    int count;
-    char *unknown = NULL;
-
-    pa_assert(fn);
-    pa_assert(str);
-    pa_assert(flags);
-
-    count = parse_list(string_conversion_table_output_flag, str, flags, &unknown);
-
-    return check_and_log(fn, ln, "flags", count, str, unknown, false);
-}
-
-#if AUDIO_API_VERSION_MAJ >= 3
-static bool parse_input_flags(const char *fn, const unsigned ln,
-                        const char *str, audio_input_flags_t *flags) {
-    int count;
-    char *unknown = NULL;
-
-    pa_assert(fn);
-    pa_assert(str);
-    pa_assert(flags);
-
-    count = parse_list(string_conversion_table_input_flag, str, flags, &unknown);
-
-    return check_and_log(fn, ln, "flags", count, str, unknown, false);
-}
-#endif
-
-static bool parse_version(const char *fn, const unsigned ln, const char *str, uint32_t *version) {
-    uint32_t version_maj;
-    uint32_t version_min;
-
-    pa_assert(fn);
-    pa_assert(str);
-    pa_assert(version);
-
-    if ((sscanf(str, "%u.%u", &version_maj, &version_min)) != 2) {
-        pa_log("[%s:%u] Failed to parse %s (%s).", fn, ln, AUDIO_HAL_VERSION_TAG, str);
-        return false;
-    } else {
-        *version = HARDWARE_DEVICE_API_VERSION(version_maj, version_min);
-        return true;
-    }
-}
-
-static void log_parse_error(const char *fn, const unsigned ln, const char *section, const char *v) {
-    pa_log("[%s:%u] failed to parse line in section %s: unknown section (%s)", fn, ln, section, v);
-}
-
-#define MAX_LINE_LENGTH (1024)
-
-bool pa_parse_droid_audio_config(const char *filename, pa_droid_config_audio *config) {
-    FILE *f;
-    unsigned n = 0;
-    bool ret = true;
-    char *full_line = NULL;
-    uint32_t hw_module_count = 0;
-
-    enum config_loc {
-        IN_ROOT             = 0,
-        IN_GLOBAL           = 1,
-        IN_GLOBAL_EXT       = 2,
-        IN_HW_MODULES       = 3,
-        IN_MODULE           = 4,
-        IN_OUTPUT_INPUT     = 5,
-        IN_CONFIG           = 6,
-        IN_MODULE_GLOBAL    = 10,
-        IN_DEVICES          = 20,
-        IN_DEVICES_DEVICE   = 21,
-        IN_GAINS            = 22,
-        IN_GAIN_N           = 23
-    } loc = IN_ROOT;
-
-    bool in_output = true;
-
-    pa_droid_config_hw_module *module = NULL;
-    pa_droid_config_output *output = NULL;
-    pa_droid_config_input *input = NULL;
-
-    pa_assert(filename);
-    pa_assert(config);
-
-    memset(config, 0, sizeof(pa_droid_config_audio));
-    config->global_config = pa_xnew0(pa_droid_config_global, 1);
-
-    f = fopen(filename, "r");
-
-    if (!f) {
-        pa_log_info("Failed to open config file (%s): %s", filename, pa_cstrerror(errno));
-        ret = false;
-        goto finish;
-    }
-
-    pa_lock_fd(fileno(f), 1);
-
-    full_line = pa_xmalloc0(sizeof(char) * MAX_LINE_LENGTH);
-
-    while (!feof(f)) {
-        char *ln, *d, *v, *value;
-
-        if (!fgets(full_line, MAX_LINE_LENGTH, f))
-            break;
-
-        n++;
-
-        pa_strip_nl(full_line);
-
-        if (!*full_line)
-            continue;
-
-        ln = full_line + strspn(full_line, WHITESPACE);
-
-        if (ln[0] == '#')
-            continue;
-
-        v = ln;
-        d = v + strcspn(v, WHITESPACE);
-
-        value = d + strspn(d, WHITESPACE);
-        d[0] = '\0';
-        d = value + strcspn(value, WHITESPACE);
-        d[0] = '\0';
-
-        /* Enter section */
-        if (pa_streq(value, "{")) {
-
-            if (!*v) {
-                pa_log("[%s:%u] failed to parse line - too few words", filename, n);
-                goto finish;
-            }
-
-            switch (loc) {
-                case IN_ROOT:
-                    if (pa_streq(v, GLOBAL_CONFIG_TAG)) {
-                        loc = IN_GLOBAL;
-                    }
-                    else if (pa_streq(v, AUDIO_HW_MODULE_TAG))
-                        loc = IN_HW_MODULES;
-                    else {
-                        log_parse_error(filename, n, "<root>", v);
-                        ret = false;
-                        goto finish;
-                    }
-                    break;
-
-                case IN_GLOBAL:
-                    if (pa_streq(v, GLOBAL_CONFIG_EXT_TAG))
-                        loc = IN_GLOBAL_EXT;
-                    else {
-                        log_parse_error(filename, n, GLOBAL_CONFIG_TAG, v);
-                        ret = false;
-                        goto finish;
-                    }
-                    break;
-
-                case IN_HW_MODULES:
-                    pa_assert(!module);
-
-                    module = pa_xnew0(pa_droid_config_hw_module, 1);
-                    SLLIST_APPEND(pa_droid_config_hw_module, config->hw_modules, module);
-                    hw_module_count++;
-                    module->name = pa_xstrndup(v, AUDIO_HARDWARE_MODULE_ID_MAX_LEN);
-                    module->config = config;
-                    loc = IN_MODULE;
-                    pa_log_debug("config: New module: %s", module->name);
-                    break;
-
-                case IN_MODULE:
-                    pa_assert(module);
-
-                    if (pa_streq(v, OUTPUTS_TAG)) {
-                        loc = IN_OUTPUT_INPUT;
-                        in_output = true;
-                    } else if (pa_streq(v, INPUTS_TAG)) {
-                        loc = IN_OUTPUT_INPUT;
-                        in_output = false;
-                    } else if (pa_streq(v, GLOBAL_CONFIG_TAG)) {
-                        loc = IN_MODULE_GLOBAL;
-                    } else if (pa_streq(v, DEVICES_TAG)) {
-                        loc = IN_DEVICES;
-                    } else {
-                        log_parse_error(filename, n, module->name, v);
-                        ret = false;
-                        goto finish;
-                    }
-                    break;
-
-                case IN_OUTPUT_INPUT:
-                    pa_assert(module);
-
-                    if (in_output) {
-                        output = pa_xnew0(pa_droid_config_output, 1);
-                        SLLIST_APPEND(pa_droid_config_output, module->outputs, output);
-                        output->name = pa_xstrndup(v, AUDIO_HARDWARE_MODULE_ID_MAX_LEN);
-                        output->module = module;
-                        loc = IN_CONFIG;
-                        pa_log_debug("config: %s: New output: %s", module->name, output->name);
-                    } else {
-                        input = pa_xnew0(pa_droid_config_input, 1);
-                        SLLIST_APPEND(pa_droid_config_input, module->inputs, input);
-                        input->name = pa_xstrndup(v, AUDIO_HARDWARE_MODULE_ID_MAX_LEN);
-                        input->module = module;
-                        loc = IN_CONFIG;
-                        pa_log_debug("config: %s: New input: %s", module->name, input->name);
-                    }
-                    break;
-
-                case IN_DEVICES:
-                    /* TODO Missing implementation of parsing the module/devices section.
-                     * As of now there is no need for the information, fix this when that
-                     * changes. */
-                    loc = IN_DEVICES_DEVICE;
-                    break;
-
-                case IN_DEVICES_DEVICE:
-                    if (pa_streq(v, GAINS_TAG))
-                        loc = IN_GAINS;
-                    else {
-                        log_parse_error(filename, n, DEVICES_TAG, v);
-                        ret = false;
-                        goto finish;
-                    }
-                    break;
-
-                case IN_GAINS:
-                    /* TODO Missing implementation of parsing the gain_n section.
-                     * As of now there is no need for the information, fix this when that
-                     * changes. */
-                    if (pa_startswith(v, GAIN_TAG_PREFIX))
-                        loc = IN_GAIN_N;
-                    else {
-                        log_parse_error(filename, n, GAINS_TAG, v);
-                        ret = false;
-                        goto finish;
-                    }
-                    break;
-
-                case IN_CONFIG:
-                    if (pa_streq(v, GAINS_TAG)) {
-                        loc = IN_GAINS;
-                    } else {
-                        log_parse_error(filename, n, in_output ? output->name : input->name, v);
-                        ret = false;
-                        goto finish;
-                    }
-                    break;
-
-                default:
-                    pa_log("[%s:%u] failed to parse line: unknown section (%s)", filename, n, v);
-                    ret = false;
-                    goto finish;
-            }
-
-            continue;
-        }
-
-        /* Exit section */
-        if (pa_streq(v, "}")) {
-            switch (loc) {
-                case IN_ROOT:
-                    pa_log("[%s:%u] failed to parse line - extra closing bracket", filename, n);
-                    ret = false;
-                    goto finish;
-
-                case IN_HW_MODULES:
-                    /* fall through */
-                case IN_GLOBAL:
-                    loc = IN_ROOT;
-                    break;
-
-                case IN_MODULE:
-                    module = NULL;
-                    loc = IN_HW_MODULES;
-                    break;
-
-                case IN_DEVICES:
-                    /* fall through */
-                case IN_MODULE_GLOBAL:
-                    loc = IN_MODULE;
-                    break;
-
-                case IN_GAINS:
-                    if (output || input)
-                        loc = IN_CONFIG;
-                    else
-                        loc = IN_DEVICES_DEVICE;
-                    break;
-
-                case IN_OUTPUT_INPUT:
-                    if (in_output)
-                        output = NULL;
-                    else
-                        input = NULL;
-                    /* fall through */
-                case IN_GAIN_N:
-                    /* fall through */
-                case IN_DEVICES_DEVICE:
-                    /* fall through */
-                case IN_CONFIG:
-                    /* fall through */
-                case IN_GLOBAL_EXT:
-                    loc--;
-                    break;
-            }
-
-            continue;
-        }
-
-        /* Parsing of values */
-        if (loc == IN_GLOBAL ||
-            loc == IN_GLOBAL_EXT ||
-            loc == IN_MODULE_GLOBAL ||
-            loc == IN_CONFIG ||
-            loc == IN_DEVICES_DEVICE ||
-            loc == IN_GAIN_N) {
-
-            bool success = false;
-
-            if (loc == IN_GLOBAL || loc == IN_MODULE_GLOBAL) {
-                pa_droid_config_global *global_config = NULL;
-
-                if (loc == IN_MODULE_GLOBAL) {
-                    pa_assert(module);
-                    if (!module->global_config)
-                        module->global_config = pa_xnew0(pa_droid_config_global, 1);
-                    global_config = module->global_config;
-                } else
-                    global_config = config->global_config;
-
-                pa_assert(global_config);
-
-                /* Parse global configuration */
-
-                if (pa_streq(v, ATTACHED_OUTPUT_DEVICES_TAG))
-                    success = parse_devices(filename, n, value, true,
-                                            &global_config->attached_output_devices, false);
-                else if (pa_streq(v, DEFAULT_OUTPUT_DEVICE_TAG))
-                    success = parse_devices(filename, n, value, true,
-                                            &global_config->default_output_device, true);
-                else if (pa_streq(v, ATTACHED_INPUT_DEVICES_TAG))
-                    success = parse_devices(filename, n, value, false,
-                                            &global_config->attached_input_devices, false);
-                else if (pa_streq(v, AUDIO_HAL_VERSION_TAG))
-                    success = parse_version(filename, n, value,
-                                            &global_config->audio_hal_version);
-#ifdef DROID_HAVE_DRC
-                // SPEAKER_DRC_ENABLED_TAG is only from Android v4.4
-                else if (pa_streq(v, SPEAKER_DRC_ENABLED_TAG))
-                    /* TODO - Add support for dynamic range control */
-                    success = true; /* Do not fail while parsing speaker_drc_enabled entry */
-#endif
-                else {
-                    pa_log("[%s:%u] failed to parse line - unknown config entry %s", filename, n, v);
-                    success = false;
-                }
-
-            } else if (loc == IN_GLOBAL_EXT) {
-
-                /* Parse custom global configuration
-                 * For now just log all custom variables, don't do
-                 * anything with the values.
-                 * TODO: Store custom values somehow */
-
-                pa_log_debug("[%s:%u] TODO custom variable: %s = %s", filename, n, v, value);
-                success = true;
-
-            } else if (loc == IN_CONFIG) {
-
-                /* Parse per-output or per-input configuration */
-
-                if ((in_output && !output) || (!in_output && !input)) {
-                    pa_log("[%s:%u] failed to parse line", filename, n);
-                    ret = false;
-                    goto finish;
-                }
-
-                if (pa_streq(v, SAMPLING_RATES_TAG))
-                    success = parse_sampling_rates(filename, n, value,
-                                                   in_output ? output->sampling_rates : input->sampling_rates);
-                else if (pa_streq(v, FORMATS_TAG))
-                    success = parse_formats(filename, n, value, in_output ? &output->formats : &input->formats);
-                else if (pa_streq(v, CHANNELS_TAG)) {
-                    if (in_output)
-                        success = parse_channels(filename, n, value, true, &output->channel_masks);
-                    else
-                        success = parse_channels(filename, n, value, false, &input->channel_masks);
-                } else if (pa_streq(v, DEVICES_TAG)) {
-                    if (in_output)
-                        success = parse_devices(filename, n, value, true, &output->devices, false);
-                    else
-                        success = parse_devices(filename, n, value, false, &input->devices, false);
-                } else if (pa_streq(v, FLAGS_TAG)) {
-                    if (in_output)
-                        success = parse_output_flags(filename, n, value, &output->flags);
-                    else {
-#if AUDIO_API_VERSION_MAJ >= 3
-                        success = parse_input_flags(filename, n, value, &input->flags);
-#else
-                        pa_log("[%s:%u] failed to parse line - output flags inside input definition", filename, n);
-                        success = false;
-#endif
-                    }
-                } else {
-                    pa_log("[%s:%u] failed to parse line - unknown config entry %s", filename, n, v);
-                    success = false;
-                }
-
-            } else if (loc == IN_DEVICES_DEVICE) {
-                /* TODO Missing implementation of parsing the module/devices section.
-                 * As of now there is no need for the information, fix this when that
-                 * changes. */
-                success = true;
-            } else if (loc == IN_GAIN_N) {
-                /* TODO Missing implementation of parsing the gain_n section.
-                 * As of now there is no need for the information, fix this when that
-                 * changes. */
-                success = true;
-            } else
-                pa_assert_not_reached();
-
-            if (!success) {
-                ret = false;
-                goto finish;
-            }
-        }
-    }
-
-    pa_log_info("Parsed config file (%s): %u modules.", filename, hw_module_count);
-
-finish:
-    if (f) {
-        pa_lock_fd(fileno(f), 0);
-        fclose(f);
-    }
-
-    pa_xfree(full_line);
-
-    return ret;
-}
-
-
-const pa_droid_config_output *pa_droid_config_find_output(const pa_droid_config_hw_module *module, const char *name) {
-    pa_droid_config_output *output;
-
-    pa_assert(module);
-    pa_assert(name);
-
-    SLLIST_FOREACH(output, module->outputs) {
-        if (pa_streq(name, output->name))
-            return output;
-    }
-
-    return NULL;
-}
-
-const pa_droid_config_input *pa_droid_config_find_input(const pa_droid_config_hw_module *module, const char *name) {
-    pa_droid_config_input *input;
-
-    pa_assert(module);
-    pa_assert(name);
-
-    SLLIST_FOREACH(input, module->inputs) {
-        if (pa_streq(name, input->name))
-            return input;
-    }
-
-    return NULL;
-}
-
-const pa_droid_config_hw_module *pa_droid_config_find_module(const pa_droid_config_audio *config, const char* module_id) {
-    pa_droid_config_hw_module *module;
-
-    pa_assert(config);
-    pa_assert(module_id);
-
-    SLLIST_FOREACH(module, config->hw_modules) {
-        if (pa_streq(module_id, module->name))
-            return module;
-    }
-
-    return NULL;
-}
 
 static pa_droid_profile *profile_new(pa_droid_profile_set *ps,
                                      const pa_droid_config_hw_module *module,
@@ -1519,18 +691,6 @@ pa_droid_mapping *pa_droid_idxset_mapping_with_device(pa_idxset *i, uint32_t dev
     return NULL;
 }
 
-bool pa_droid_output_port_name(audio_devices_t value, const char **to_str) {
-    return string_convert_num_to_str(string_conversion_table_output_device_fancy, (uint32_t) value, to_str);
-}
-
-bool pa_droid_input_port_name(audio_devices_t value, const char **to_str) {
-    return string_convert_num_to_str(string_conversion_table_input_device_fancy, (uint32_t) value, to_str);
-}
-
-bool pa_droid_audio_source_name(audio_source_t value, const char **to_str) {
-    return string_convert_num_to_str(string_conversion_table_audio_source_fancy, (uint32_t) value, to_str);
-}
-
 static int add_ports(pa_core *core, pa_card_profile *cp, pa_hashmap *ports, pa_droid_mapping *am, pa_hashmap *extra) {
     pa_droid_port *p;
     pa_device_port_new_data dp_data;
@@ -1973,7 +1133,7 @@ static void droid_hw_module_close(pa_droid_hw_module *hw) {
         pa_hook_done(&hw->hooks[h]);
 
     if (hw->config)
-        droid_config_free(hw->config);
+        pa_droid_config_free(hw->config);
 
     if (hw->device && !pa_droid_quirk(hw, QUIRK_UNLOAD_NO_CLOSE))
         audio_hw_device_close(hw->device);
@@ -2015,72 +1175,6 @@ void pa_droid_hw_module_unref(pa_droid_hw_module *hw) {
 
     pa_assert_se(pa_shared_remove(hw->core, hw->shared_name) >= 0);
     droid_hw_module_close(hw);
-}
-
-static void droid_config_free(pa_droid_config_audio *config) {
-    pa_droid_config_hw_module *module;
-    pa_droid_config_output *output;
-    pa_droid_config_input *input;
-
-    pa_assert(config);
-
-    while (config->hw_modules) {
-        SLLIST_STEAL_FIRST(module, config->hw_modules);
-
-        while (module->outputs) {
-            SLLIST_STEAL_FIRST(output, module->outputs);
-            pa_xfree(output->name);
-            pa_xfree(output);
-        }
-
-        while (module->inputs) {
-            SLLIST_STEAL_FIRST(input, module->inputs);
-            pa_xfree(input->name);
-            pa_xfree(input);
-        }
-
-        pa_xfree(module->global_config);
-        pa_xfree(module->name);
-        pa_xfree(module);
-    }
-
-    pa_xfree(config->global_config);
-    pa_xfree(config);
-}
-
-pa_droid_config_audio *pa_droid_config_load(pa_modargs *ma) {
-    pa_droid_config_audio *config;
-    const char *config_location;
-
-    pa_assert(ma);
-
-    config = pa_xnew0(pa_droid_config_audio, 1);
-
-    if ((config_location = pa_modargs_get_value(ma, "config", NULL))) {
-        if (!pa_parse_droid_audio_config(config_location, config)) {
-            pa_log("Failed to parse configuration from %s", config_location);
-            goto fail;
-        }
-    } else {
-        config_location = AUDIO_POLICY_VENDOR_CONFIG_FILE;
-
-        if (!pa_parse_droid_audio_config(config_location, config)) {
-            pa_log_debug("Failed to parse configuration from vendor %s", config_location);
-
-            config_location = AUDIO_POLICY_CONFIG_FILE;
-
-            if (!pa_parse_droid_audio_config(config_location, config)) {
-                pa_log("Failed to parse configuration from %s", config_location);
-                goto fail;
-            }
-        }
-    }
-
-    return config;
-
-fail:
-    droid_config_free(config);
-    return NULL;
 }
 
 void pa_droid_hw_module_lock(pa_droid_hw_module *hw) {
